+++ upidashf/public/index.html
@@ -29,8 +29,7 @@
 
 <body>
   <noscript>You need to enable JavaScript to run this app.</noscript>
-  <div id="root" style="background-image=url('./background.jpg'); background-repeat: no-repeat='no-repeat'; bakcground.size='cover'
-    , margin-bottom=10px">
+  <div id="root">
   </div>
   <!--
       This HTML file is a template.
@@ -44,18 +43,6 @@
     -->
 
   <script>
-    // let descopeSdk = Descope({ projectId: "P2TsKReJyvvtsjHQ5DByI3N9Fckr" });
-    async function f1() {
-      await descopeSdk.me().then((res) => { console.log(res) }).catch((err) => {
-        console.log(err);
-        // async function f2() {
-        //   await descopeSdk.refresh()
-        // }
-        // f2()
-
-      });
-    }
-    f1();
   </script>
 </body>
